Added a proposed solution for the tictactoe game
