--- files/ttt_closed/tictactoe.cpp
@@ -1,5 +1,0 @@
-// This is awefull. But we need a single .cpp-file and this way we can change
-// the implementation at all points at the same time
-#include "frage_feld_nummer.cpp"
-#include "gebe_feld_aus.cpp"
-#include "gewinnerin.cpp"
